Attempt at some cleanup
